# Click-Reserva/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClickReserva API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /professors:
    get:
      operationId: getProfessors
      tags: [professors]
      summary: List all professors
      responses:
        "200":
          description: List of professors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Professor"
    post:
      operationId: createProfessor
      tags: [professors]
      summary: Create a professor (coordinator only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfessorBody"
      responses:
        "201":
          description: Professor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professor"
  /professors/{professorId}:
    get:
      operationId: getProfessor
      tags: [professors]
      summary: Get a single professor
      parameters:
        - name: professorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Professor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professor"
    put:
      operationId: updateProfessor
      tags: [professors]
      summary: Update a professor
      parameters:
        - name: professorId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfessorBody"
      responses:
        "200":
          description: Professor updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professor"
  /professors/{professorId}/unblock:
    post:
      operationId: unblockProfessor
      tags: [professors]
      summary: Unblock a professor (coordinator only)
      parameters:
        - name: professorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Professor unblocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Professor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [professor, coordinator]
        blocked:
          type: boolean
        totalAbsences:
          type: integer
        blockedAt:
          type: string
          nullable: true
        blockReason:
          type: string
          nullable: true
        isActive:
          type: boolean
      required:
        - id
        - name
        - email
        - role
        - blocked
        - totalAbsences
        - isActive
    CreateProfessorBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [professor, coordinator]
      required:
        - name
        - email
        - password
    UpdateProfessorBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        isActive:
          type: boolean
...